Fix 32-bit compatibility with int64 casts for file times

+++ utils/file_times_darwin.go
@@ -9,5 +9,6 @@ import (
 
 // getBirthTime returns the birth time (creation time) on macOS.
 func getBirthTime(stat *syscall.Stat_t) time.Time {
-	return time.Unix(stat.Birthtimespec.Sec, stat.Birthtimespec.Nsec)
+	// Cast to int64 for consistency and 32-bit compatibility
+	return time.Unix(int64(stat.Birthtimespec.Sec), int64(stat.Birthtimespec.Nsec))
 }
